Fix an OBE in the Model vertex, texture, and normal indices.
A couple small modifications to the OBJFile class to handle differences in the 3DS max .obj file.

A couple vertex shaders used to test normal and texture coordinates.

Added texture coordinate support to faces and models.

Changed (fixed) how faces use the VBO.

Put a textured happy fun ball (do not taunt!) in the scene. UV coords are a bit wacky, but not terrible for basic unwrapping.